# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Anomaly Detection API",
    "description": "Backend foundation for Machine Learning Enhanced Graph-Based Anomaly Detection for Financial Transactions. Provides transaction ingestion, graph construction, and behavioral/temporal feature extraction.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transactions/upload": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Upload Transactions CSV",
        "description": "Uploads a CSV file containing transaction data. Validates canonical columns (transaction_id, sender_id, receiver_id, amount, timestamp), rejects invalid records without silent corruption, creates a dataset-scoped state, constructs the transaction graph, and extracts behavioral/temporal features.",
        "operationId": "upload_transactions_api_v1_transactions_upload_post",
        "parameters": [
          {
            "name": "strict_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If True, any single invalid row aborts the entire dataset upload.",
              "default": false,
              "title": "Strict Mode"
            },
            "description": "If True, any single invalid row aborts the entire dataset upload."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_transactions_api_v1_transactions_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{dataset_id}/status": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Dataset Processing Status",
        "description": "Check processing status of a dataset by its unique dataset ID.",
        "operationId": "get_dataset_status_api_v1_transactions__dataset_id__status_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{dataset_id}/summary": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Transaction Statistics Summary",
        "description": "Returns aggregate statistical metrics for the specified dataset.",
        "operationId": "get_transaction_summary_api_v1_transactions__dataset_id__summary_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_transactions_api_v1_transactions_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CSV file containing transactions"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_transactions_api_v1_transactions_upload_post"
      },
...
      "DatasetStatusResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "status"
        ],
        "title": "DatasetStatusResponse",
        "description": "Status response for dataset processing/availability."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TimeRange": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          }
        },
        "type": "object",
        "title": "TimeRange"
      },
      "TransactionSummaryResponse": {
        "properties": {
          "transactions": {
            "type": "integer",
            "title": "Transactions",
            "description": "Total count of valid transactions"
          },
          "users": {
            "type": "integer",
            "title": "Users",
            "description": "Total count of distinct entities/users"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount",
            "description": "Total cumulative monetary volume"
          },
          "average_amount": {
            "type": "number",
            "title": "Average Amount",
            "description": "Average transaction amount"
          },
          "min_amount": {
            "type": "number",
            "title": "Min Amount",
            "description": "Minimum transaction amount"
          },
          "max_amount": {
            "type": "number",
            "title": "Max Amount",
            "description": "Maximum transaction amount"
          },
          "time_range": {
            "$ref": "#/components/schemas/TimeRange"
          }
        },
        "type": "object",
        "required": [
          "transactions",
          "users",
          "total_amount",
          "average_amount",
          "min_amount",
          "max_amount"
        ],
        "title": "TransactionSummaryResponse",
        "description": "Summary of ingested transactions."
      },
      "TransactionUploadResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "complete"
          },
          "total_rows_parsed": {
            "type": "integer",
            "title": "Total Rows Parsed"
          },
          "valid_transactions": {
            "type": "integer",
            "title": "Valid Transactions"
          },
          "invalid_rows_count": {
            "type": "integer",
            "title": "Invalid Rows Count"
          },
          "validation_errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Validation Errors",
            "default": []
          },
          "duplicate_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Duplicate Warnings",
            "default": []
          },
          "summary": {
            "$ref": "#/components/schemas/TransactionSummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "message",
          "dataset_id",
          "total_rows_parsed",
          "valid_transactions",
          "invalid_rows_count",
          "summary"
        ],
        "title": "TransactionUploadResponse",
        "description": "Response returned upon CSV transaction file upload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}